Split routes/datasets/crud.py into listings/status/staging/load + _helpers
The 970-LOC datasets/crud.py mixed four concerns (read-only listings,
status/progress/cancel, importer staging + dispatch, and the various
load endpoints). Split it along the seams called out in
docs/plans/codebase-reorg.md #2:

- listings.py        (~115 LOC) — /api/media-types, /api/embedders,
                                  /api/clippers, /api/converters,
                                  /api/dataset/{importers,all-importers}
- status.py          (~80 LOC)  — /api/dataset/{status,progress,
                                  loading-tasks,cancel,cancel/<id>}
- staging.py         (~220 LOC) — /api/dataset/{available-files,combine,
                                  stage-*,staging,import/<n>,
                                  import/<n>/options}
- load.py            (~330 LOC) — /api/dataset/{load-*,import-local-folder,
                                  export,clear}
- _helpers.py        (~110 LOC) — _normalize_media_type_param,
                                  _extract_importer_fields,
                                  _extract_clipper_params,
                                  _safe_relative_upload_path

Each file owns its own Flask Blueprint, registered directly on the app
(matching the existing routes/detectors/ pattern). The old
``datasets_bp`` re-export shim is gone; ``datasets_ui_bp`` is now
registered at the app level instead of nested under ``datasets_bp``.

Test patches that targeted ``vtsearch.routes.datasets.crud.<sym>`` are
updated to point at the new owning module (the staging/load split
follows which endpoint the test hits). Symbols that always lived in
``vtsearch.datasets.load_pipeline`` are now patched at their source.

No URL or behaviour changes.
