+++ tests/datasets/test_combine_datasets.py
@@ -459,7 +459,7 @@ class TestCombineEndpoint:
         _write_pickle_dataset(p2, ds2)
 
         # Mock the background loader to prevent it from clearing global medias
-        with patch("vtsearch.routes.datasets.crud._run_importer_in_background"):
+        with patch("vtsearch.routes.datasets.staging._run_importer_in_background"):
             resp = client.post(
                 "/api/dataset/combine",
                 json={"datasets": [str(p1), str(p2)]},
@@ -597,7 +597,7 @@ class TestClearStagingEndpoint:
 
     def test_clear_staging_removes_files(self, client):
         """Staging files are actually removed."""
-        from vtsearch.routes.datasets.crud import STAGING_DIR
+        from vtsearch.datasets.load_pipeline import STAGING_DIR
 
         STAGING_DIR.mkdir(parents=True, exist_ok=True)
         test_file = STAGING_DIR / "stage_test.pkl"
